@@ -1,16 +1,20 @@
 package com.thinkup.sampleeasy
 
-import android.view.View
+import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.thinkup.easycore.ViewRenderer
-import kotlinx.android.synthetic.main.item_sample.view.*
+import com.thinkup.sampleeasy.databinding.ItemSampleBinding
 
-class SampleRenderer : ViewRenderer<SampleItem, View>(SampleItem::class) {
-    override fun create(parent: ViewGroup): View = inflate(R.layout.item_sample, parent, false)
+class SampleRenderer : ViewRenderer<SampleItem, ItemSampleBinding>(SampleItem::class) {
 
-    override fun bind(view: View, model: SampleItem, position: Int) {
-        view.sampleName.text = model.name
-        view.sampleGender.text = model.gender
-        view.sampleCompany.text = model.company
+    override val create: (inflater: LayoutInflater, parent: ViewGroup, attachToParent: Boolean) -> ItemSampleBinding
+        get() = ItemSampleBinding::inflate
+
+    override fun bind(binding: ItemSampleBinding, model: SampleItem, position: Int) {
+        with(binding) {
+            sampleName.text = model.name
+            sampleGender.text = model.gender
+            sampleCompany.text = model.company
+        }
     }
 }
